feat : main page works
