+++ src/app/works/page.module.scss
@@ -15,15 +15,4 @@
             font-size: 16px;
         }
     }
-
-    .works{
-        margin: 16px 0;
-    }
-
-    .loaderContainer{
-        margin: 32px 0;
-        display: flex;
-        justify-content: center;
-        align-items: center;
-    }
 }
