+++ src/components/features/Works/index.module.scss
@@ -1,0 +1,19 @@
+.container{
+
+    .works{
+        margin: 16px 0;
+    }
+
+    .loaderContainer{
+        margin: 32px 0;
+        display: flex;
+        justify-content: center;
+        align-items: center;
+    }
+
+    .buttonContainer {
+        display: flex;
+        justify-content: center;
+        align-items: center;
+    }
+}
